refactor folder names
