+++ src/gatling/simulations/utils/Environment.scala
@@ -1,4 +1,4 @@
-package util
+package utils
 
 object Environment {
